Improve tracking of player changes
The old logic failed when a player's turn ended without requiring player input.
This player's name could not be inferred from the previous game state, nor from the next.
This caused confusing in the clients, in particular, an assertion failure in PICO-8.

Changes are now handled by letting the backend signal them when needed.